Enforce consistent config handling for services
This commit makes the sensu-backend and sensu-agent services consistent
in their expectations and handling of the /etc/sensu/{backend,agent}.yml
config files.

With systemd, the ConditionPathExists feature has been applied so that
if the configuration file is missing, the service will not start up.

sensu-agent has been modified to use an explicit path to a configuration
file. This should prevent issues with invalid configuration. (See
sensu-packaging #9) This has been done in both systemd and sysvinit.

Signed-off-by: Eric Chlebek <eric@sensu.io>